+++ stripe_terminal/ios/Classes/Utils.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+func createApiException(_ code: TerminalExceptionCodeApi, _ message: String? = nil) -> TerminalExceptionApi {
+    return TerminalExceptionApi(
+        code: code,
+        message: message ?? "",
+        stackTrace: nil,
+        paymentIntent: nil,
+        apiError: nil
+    )
+}